Add jenga command for performance testing
