+++ common/src/main/kotlin/org/valkyrienskies/mod/common/command/VSCommands.kt
@@ -24,7 +24,7 @@ import org.valkyrienskies.mod.common.command.commands.RemassCommand
 import org.valkyrienskies.mod.common.command.commands.RenameCommand
 import org.valkyrienskies.mod.common.command.commands.ScaleCommand
 import org.valkyrienskies.mod.common.command.commands.SplittingCommand
-import org.valkyrienskies.mod.common.command.commands.SpawnShipCubeCommand
+import org.valkyrienskies.mod.common.command.commands.PerfTestCommand
 import org.valkyrienskies.mod.common.command.commands.StaticCommand
 import org.valkyrienskies.mod.common.command.commands.TeleportCommand
 import org.valkyrienskies.mod.common.shipObjectWorld
@@ -48,7 +48,7 @@ object VSCommands {
         RenameCommand.register(vs)
         ScaleCommand.register(vs)
         SplittingCommand.register(vs)
-        SpawnShipCubeCommand.register(vs)
+        PerfTestCommand.register(vs)
         StaticCommand.register(vs)
         TeleportCommand.register(vs)
